Added support for IgnoredWarnings
Renamed the previous ExpectedWarnings to IgnoredWarnings and added ExpectedWarnings.
ExpectedWarnings: Warnings like #warning and [Obsolete] that should display but not fail the build.
IgnoredWarnings: Warnings that we are leaving the code for elucidation or because the correct code to remove them has not been covered yet.  They are ignored because we don't want them to fail the build.